pyproject.toml
[tool.poetry]
name = "textsummarizationandkeywordextraction"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
asgiref = "3.3.4"
colorama = "0.4.4"
cycler = "0.10.0"
typing-extensions = "3.10.0.0"
joblib = "1.0.1"
importlib-metadata = "4.0.1"
zipp = "3.4.1"
tqdm = "4.60.0"
regex = "2021.4.4"
numpy = "1.20.3"
pandas = "1.2.4"
python-dateutil = "2.8.1"
pytz = "2021.1"
kiwisolver = "1.3.1"
sqlparse = "0.4.1"
matplotlib = "3.4.2"
Pillow = "8.2.0"
click = "8.0.1"
pyparsing = "2.4.7"
Django = "2.2.4"
nltk = "3.6.2"
six = "1.16.0"
newspaper3k = "^0.2.8"
gensim = "3.8.3"
pytesseract = "^0.3.9"
opencv-python = "4.2.0.32"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
